+++ push_notifications/gcm.py
@@ -1,25 +1,37 @@
 """
-Google Cloud Messaging
-Previously known as C2DM
-Documentation is available on the Android Developer website:
-https://developer.android.com/google/gcm/index.html
+Firebase Cloud Messaging
+Previously known as GCM / C2DM
+Documentation is available on the Firebase Developer website:
+https://firebase.google.com/docs/cloud-messaging/
 """
 
 import json
 from .models import GCMDevice
 
-
 try:
 	from urllib.request import Request, urlopen
-	from urllib.parse import urlencode
 except ImportError:
 	# Python 2 support
 	from urllib2 import Request, urlopen
-	from urllib import urlencode
 
 from django.core.exceptions import ImproperlyConfigured
 from . import NotificationError
 from .settings import PUSH_NOTIFICATIONS_SETTINGS as SETTINGS
+
+
+# Valid keys for FCM messages
+# See ref : https://firebase.google.com/docs/cloud-messaging/http-server-ref#notification-payload-support
+FCM_TARGETS_KEYS = [
+	'to', 'condition', 'notification_key'
+]
+FCM_OPTIONS_KEYS = [
+	'collapse_key', 'priority', 'content_available', 'delay_while_idle', 'time_to_live',
+	'restricted_package_name', 'dry_run'
+]
+FCM_NOTIFICATIONS_PAYLOAD_KEYS = [
+	'title', 'body', 'icon', 'sound', 'badge', 'color', 'tag', 'click_action',
+	'body_loc_key', 'body_loc_args', 'title_loc_key', 'title_loc_args'
+]
 
 
 class GCMError(NotificationError):
@@ -34,7 +46,7 @@ def _chunks(l, n):
 		yield l[i:i + n]
 
 
-def _gcm_send(data, content_type):
+def _gcm_send(payload, content_type):
 	key = SETTINGS.get("GCM_API_KEY")
 	if not key:
 		raise ImproperlyConfigured(
@@ -44,13 +56,13 @@ def _gcm_send(data, content_type):
 	headers = {
 		"Content-Type": content_type,
 		"Authorization": "key=%s" % (key),
-		"Content-Length": str(len(data)),
+		"Content-Length": str(len(payload)),
 	}
-	request = Request(SETTINGS["GCM_POST_URL"], data, headers)
+	request = Request(SETTINGS["GCM_POST_URL"], payload, headers)
 	return urlopen(request, timeout=SETTINGS["GCM_ERROR_TIMEOUT"]).read().decode("utf-8")
 
 
-def _fcm_send(data, content_type):
+def _fcm_send(payload, content_type):
 	key = SETTINGS.get("FCM_API_KEY")
 	if not key:
 		raise ImproperlyConfigured(
@@ -60,71 +72,13 @@ def _fcm_send(data, content_type):
 	headers = {
 		"Content-Type": content_type,
 		"Authorization": "key=%s" % (key),
-		"Content-Length": str(len(data)),
+		"Content-Length": str(len(payload)),
 	}
-	request = Request(SETTINGS["FCM_POST_URL"], data, headers)
+	request = Request(SETTINGS["FCM_POST_URL"], payload, headers)
 	return urlopen(request, timeout=SETTINGS["FCM_ERROR_TIMEOUT"]).read().decode("utf-8")
 
 
-def _cm_send_plain(registration_id, data, cloud_type="GCM", **kwargs):
-	"""
-	Sends a GCM notification to a single registration_id or to a
-	topic (If "topic" included in the kwargs).
-	This will send the notification as form data.
-	If sending multiple notifications, it is more efficient to use
-	gcm_send_bulk_message() with a list of registration_ids
-	"""
-
-	values = {"registration_id": registration_id} if registration_id else {}
-
-	for k, v in data.items():
-		values["data.%s" % (k)] = v.encode("utf-8")
-
-	for k, v in kwargs.items():
-		if v:
-			if isinstance(v, bool):
-				# Encode bools into ints
-				v = 1
-			values[k] = v
-
-	data = urlencode(sorted(values.items())).encode("utf-8")  # sorted items for tests
-	if cloud_type == "GCM":
-		send_func = _gcm_send
-	elif cloud_type == "FCM":
-		send_func = _fcm_send
-	else:
-		raise ImproperlyConfigured("cloud_type must be GCM or FCM not %r" % (cloud_type))
-	result = send_func(data, "application/x-www-form-urlencoded;charset=UTF-8")
-
-	# Information about handling response from Google docs
-	# (https://developers.google.com/cloud-messaging/http):
-	# If first line starts with id, check second line:
-	# If second line starts with registration_id, gets its
-	# value and replace the registration tokens in your
-	# server database. Otherwise, get the value of Error
-
-	if result.startswith("id"):
-		lines = result.split("\n")
-		if len(lines) > 1 and lines[1].startswith("registration_id"):
-			new_id = lines[1].split("=")[-1]
-			_gcm_handle_canonical_id(new_id, registration_id, cloud_type)
-
-	elif result.startswith("Error="):
-		if result in ("Error=NotRegistered", "Error=InvalidRegistration"):
-			# Deactivate the problematic device
-			device = GCMDevice.objects.filter(
-				registration_id=values["registration_id"],
-				cloud_message_type=cloud_type,
-			)
-			device.update(active=0)
-			return result
-
-		raise GCMError(result)
-
-	return result
-
-
-def _handler_cm_message_json(registration_ids, response_data, cloud_type):
+def _cm_handle_response(registration_ids, response_data, cloud_type):
 	response = response_data
 	if response.get("failure") or response.get("canonical_ids"):
 		ids_to_remove, old_new_ids = [], []
@@ -132,22 +86,17 @@ def _handler_cm_message_json(registration_ids, response_data, cloud_type):
 		for index, result in enumerate(response["results"]):
 			error = result.get("error")
 			if error:
-				# Information from Google docs
-				# https://developers.google.com/cloud-messaging/http
-				# If error is NotRegistered or InvalidRegistration,
-				# then we will deactivate devices because this
-				# registration ID is no more valid and can't be used
-				# to send messages, otherwise raise error
+				# https://firebase.google.com/docs/cloud-messaging/http-server-ref#error-codes
+				# If error is NotRegistered or InvalidRegistration, then we will deactivate devices because
+				# this registration ID is no more valid and can't be used to send messages, otherwise raise error
 				if error in ("NotRegistered", "InvalidRegistration"):
 					ids_to_remove.append(registration_ids[index])
 				else:
 					throw_error = True
 
-			# If registration_id is set, replace the original ID with
-			# the new value (canonical ID) in your server database.
-			# Note that the original ID is not part of the result, so
-			# you need to obtain it from the list of registration_ids
-			# passed in the request (using the same index).
+			# If registration_id is set, replace the original ID with the new value (canonical ID) in your
+			# server database. Note that the original ID is not part of the result, so you need to obtain it
+			# from the list of registration_ids passed in the request (using the same index).
 			new_id = result.get("registration_id")
 			if new_id:
 				old_new_ids.append((registration_ids[index], new_id))
@@ -157,43 +106,61 @@ def _handler_cm_message_json(registration_ids, response_data, cloud_type):
 			removed.update(active=0)
 
 		for old_id, new_id in old_new_ids:
-			_gcm_handle_canonical_id(new_id, old_id, cloud_type)
+			_cm_handle_canonical_id(new_id, old_id, cloud_type)
 
 		if throw_error:
 			raise GCMError(response)
 	return response
 
 
-def _cm_send_json(registration_ids, data, cloud_type="GCM", **kwargs):
+def _cm_send_request(registration_ids, data, cloud_type="GCM", use_fcm_notifications=True, **kwargs):
 	"""
-	Sends a GCM notification to one or more registration_ids. The registration_ids
-	needs to be a list.
-	This will send the notification as json data.
+	Sends a FCM or GCM notification to one or more registration_ids as json data.
+	The registration_ids needs to be a list.
 	"""
 
-	values = {"registration_ids": registration_ids} if registration_ids else {}
+	payload = {"registration_ids": registration_ids} if registration_ids else {}
 
-	if data is not None:
-		values["data"] = data
+	# If using FCM, optionnally autodiscovers notification related keys
+	# https://firebase.google.com/docs/cloud-messaging/concept-options#notifications_and_data_messages
+	if cloud_type == "FCM" and use_fcm_notifications:
+		notification_payload = {}
+		if 'message' in data:
+			notification_payload['body'] = data.pop('message', None)
 
-	for k, v in kwargs.items():
-		if v:
-			values[k] = v
+		for key in FCM_NOTIFICATIONS_PAYLOAD_KEYS:
+			value_from_extra = data.pop(key, None)
+			if value_from_extra:
+				notification_payload[key] = value_from_extra
+			value_from_kwargs = kwargs.pop(key, None)
+			if value_from_kwargs:
+				notification_payload[key] = value_from_kwargs
+		if notification_payload:
+			payload['notification'] = notification_payload
+
+	if data:
+		payload['data'] = data
+
+	# Attach any additional non falsy keyword args (targets, options)
+	# See ref : https://firebase.google.com/docs/cloud-messaging/http-server-ref#table1
+	payload.update({k: v for k, v in kwargs.items() if v and (k in FCM_TARGETS_KEYS or k in FCM_OPTIONS_KEYS)})
 
 	# Sort the keys for deterministic output (useful for tests)
-	data = json.dumps(values, separators=(",", ":"), sort_keys=True).encode("utf-8")
+	json_payload = json.dumps(payload, separators=(",", ":"), sort_keys=True).encode("utf-8")
+
+	# Sends requests and handles the response
 	if cloud_type == "GCM":
-		response = json.loads(_gcm_send(data, "application/json"))
+		response = json.loads(_gcm_send(json_payload, "application/json"))
 	elif cloud_type == "FCM":
-		response = json.loads(_fcm_send(data, "application/json"))
+		response = json.loads(_fcm_send(json_payload, "application/json"))
 	else:
-		raise ImproperlyConfigured("cloud_type must be GCM or FCM not %s" % str(cloud_type))
-	return _handler_cm_message_json(registration_ids, response, cloud_type)
+		raise ImproperlyConfigured("cloud_type must be FCM or GCM not %s" % str(cloud_type))
+	return _cm_handle_response(registration_ids, response, cloud_type)
 
 
-def _gcm_handle_canonical_id(canonical_id, current_id, cloud_type):
+def _cm_handle_canonical_id(canonical_id, current_id, cloud_type):
 	"""
-	Handle situation when GCM server response contains canonical ID
+	Handle situation when FCM server response contains canonical ID
 	"""
 	if GCMDevice.objects.filter(registration_id=canonical_id, cloud_message_type=cloud_type, active=True).exists():
 		GCMDevice.objects.filter(registration_id=current_id, cloud_message_type=cloud_type).update(active=False)
@@ -202,46 +169,36 @@ def _gcm_handle_canonical_id(canonical_id, current_id, cloud_type):
 			.update(registration_id=canonical_id)
 
 
-def send_message(registration_id, data, cloud_type, **kwargs):
+def send_message(registration_ids, data, cloud_type, **kwargs):
 	"""
-	Sends a GCM or FCM notification to a single registration_id.
-
-	If sending multiple notifications, it is more efficient to use
-	send_bulk_message() with a list of registration_ids
+	Sends a FCM (or GCM) notification to one or more registration_ids. The registration_ids
+	can be a list or a single string. This will send the notification as json data.
 
 	A reference of extra keyword arguments sent to the server is available here:
-	https://developers.google.com/cloud-messaging/server-ref#downstream
+	https://firebase.google.com/docs/cloud-messaging/http-server-ref#table1
 	"""
-
-	if registration_id:
-		return _cm_send_plain(registration_id, data, cloud_type, **kwargs)
-
-
-def send_bulk_message(registration_ids, data, cloud_type, **kwargs):
-	"""
-	Sends a GCM or FCM notification to one or more registration_ids. The registration_ids
-	needs to be a list.
-	This will send the notification as json data.
-
-	A reference of extra keyword arguments sent to the server is available here:
-	https://firebase.google.com/docs/cloud-messaging/send-message
-	"""
-	if cloud_type == "GCM":
-		max_recipients = SETTINGS.get("GCM_MAX_RECIPIENTS")
-	elif cloud_type == "FCM":
+	if cloud_type == "FCM":
 		max_recipients = SETTINGS.get("FCM_MAX_RECIPIENTS")
+	elif cloud_type == "GCM":
+		max_recipients = SETTINGS.get("GCM_MAX_RECIPIENTS")
 	else:
-		raise ImproperlyConfigured("cloud_type must be GCM or FCM not %s" % str(cloud_type))
+		raise ImproperlyConfigured("cloud_type must be FCM or GCM not %s" % str(cloud_type))
 
+	# Checks for valid recipient
 	if registration_ids is None and "/topics/" not in kwargs.get("to", ""):
 		return
-	# GCM only allows up to 1000 reg ids per bulk message
-	# https://developer.android.com/google/gcm/gcm.html#request
-	if registration_ids:
-		if len(registration_ids) > max_recipients:
-			ret = []
-			for chunk in _chunks(registration_ids, max_recipients):
-				ret.append(_cm_send_json(chunk, data, cloud_type=cloud_type, **kwargs))
-			return ret
 
-	return _cm_send_json(registration_ids, data, cloud_type=cloud_type, **kwargs)
+	# Bundles the registration_ids in an list if only one is sent
+	if not isinstance(registration_ids, list):
+		registration_ids = [registration_ids] if registration_ids else None
+
+	# FCM only allows up to 1000 reg ids per bulk message
+	# https://firebase.google.com/docs/cloud-messaging/server#http-request
+	if registration_ids:
+		ret = []
+		for chunk in _chunks(registration_ids, max_recipients):
+			ret.append(_cm_send_request(chunk, data, cloud_type=cloud_type, **kwargs))
+		return ret[0] if len(ret) == 1 else ret
+
+
+send_bulk_message = send_message